+++ src/youwol/app/environment/models/models_config.py
@@ -37,6 +37,7 @@ from .model_remote import (
     CloudEnvironments,
     Connection,
 )
+from .models_features import Features
 from .models_project import Projects
 from .models_token_storage import TokensStorageConf, TokensStorageSystemKeyring
 
@@ -485,3 +486,5 @@ class Configuration(BaseModel):
     """
     Various handles for customization (*e.g.* middleware, commands, ...)
     """
+
+    features: Features = Features()